cambios de validaciones de relaciones de consola/genero a video juegos

+++ VideoJuegoAPI/Controllers/ConsolasController.cs
@@ -111,14 +111,35 @@ namespace VideoJuegoAPI.Controllers
             ResposeObject response = new ResposeObject();
             try
             {
-                Consola consola = _context.Consolas.Find(id);
-                _context.Consolas.Remove(consola);
-                await _context.SaveChangesAsync();
+                VideoJuego video;
 
-                response.Status = 1;
-                response.Message = "Se ejecuto correctamente";
+                try
+                {
+                    video = _context.VideoJuegos.Where(c => c.IdConsola == id).Single();
+                }
+                catch { 
+                    video = null;
+                }
 
-                return StatusCode(StatusCodes.Status200OK, response);
+                if (video != null)
+                {
+                    response.Status = 0;
+                    response.Message = "No es posible eliminar la consola, contiene videojuegos relacionados ";
+
+                    return StatusCode(StatusCodes.Status400BadRequest, response);
+                }
+                else
+                {
+                    Consola consola = _context.Consolas.Find(id);
+                    _context.Consolas.Remove(consola);
+                    await _context.SaveChangesAsync();
+
+                    response.Status = 1;
+                    response.Message = "Se ejecuto correctamente";
+
+                    return StatusCode(StatusCodes.Status200OK, response);
+                }
+
             }
             catch (Exception ex)
             {
